+++ apiserver/paasng/paasng/accessories/publish/sync_market/constant.py
@@ -17,24 +17,12 @@ We undertake not to change the open source license (MIT license) applicable
 to the current version of the project delivered to anyone in the future.
 """
 from dataclasses import dataclass
-from typing import List, Optional
+from typing import List
 
+from paasng.anti_contracts.accessories.publish.sync_market import constants
 from paasng.utils.basic import ChoicesEnum
 
-
-class RegionConverter:
-    _data = {
-        "ieod": "ied",
-    }
-    _data_reverse = {v: k for k, v in list(_data.items())}
-
-    @classmethod
-    def to_old(cls, region):
-        return cls._data.get(region, region)
-
-    @classmethod
-    def to_new(cls, region):
-        return cls._data_reverse.get(region, region)
+RegionConverter = constants.RegionConverter
 
 
 class PVTimeType(ChoicesEnum):
@@ -65,13 +53,7 @@ class PVTimeType(ChoicesEnum):
     )
 
 
-@dataclass
-class EnvItem:
-    key: str
-    value: str
-    description: str
-    is_builtin: bool
-    environment_name: Optional[str] = None
+EnvItem = constants.EnvItem
 
 
 @dataclass
