+++ apiserver/paasng/paasng/platform/mgrlegacy/constants.py
@@ -18,6 +18,7 @@ to the current version of the project delivered to anyone in the future.
 """
 from dataclasses import dataclass
 
+from paasng.anti_contracts.platform.mgrlegacy import constants
 from paasng.utils.basic import ChoicesEnum
 
 
@@ -87,26 +88,7 @@ class LegacyAppTag(ChoicesEnum):
     _choices_labels = ((SUPPORT, u'支持'), (NOT_SUPPORT, u'不支持'), (ON_MIGRATION, u'迁移中'), (FINISHED_MIGRATION, u'迁移完成'))
 
 
-class LegacyAppState(ChoicesEnum):
-    """命名保持跟 PaaS2.0 一致，方便核对"""
-
-    OUTLINE = 0
-    DEVELOPMENT = 1
-    TEST = 3
-    ONLINE = 4
-    IN_TEST = 8
-    IN_ONLINE = 9
-    IN_OUTLINE = 10
-
-    _choices_labels = (
-        (OUTLINE, '已下架'),
-        (DEVELOPMENT, '开发中'),
-        (TEST, '测试中'),
-        (ONLINE, '已上线'),
-        (IN_TEST, '正在提测'),
-        (IN_ONLINE, '正在上线'),
-        (IN_OUTLINE, '正在下架'),
-    )
+LegacyAppState = constants.LegacyAppState
 
 
 @dataclass
